# RAG: Complete knowledge base implementation with Docling, semantic search, and Redis caching