--- a/backend/app/routers/__init__.py
+++ b/backend/app/routers/__init__.py
@@ -2,10 +2,11 @@
 API routers for FastAPI endpoints
 """
 
-from app.routers import repositories, technologies, dashboard
+from app.routers import repositories, technologies, dashboard, knowledge
 
 __all__ = [
     "repositories",
     "technologies",
     "dashboard",
+    "knowledge",
 ]